move getting start page to main page, and main page to template page, reorder nav bar, fix accessability, improve ui, clean up
